+++ packages/ui/react-ui-form/src/components/Form/Form.theme.ts
@@ -19,8 +19,6 @@ const formStyles = tv({
     sectionDescription: 'text-description',
     fieldSet: '',
     field: '',
-    labelContainer: '',
-    labelText: '',
     description: 'text-description',
     control: '',
     validation: '',
@@ -35,8 +33,6 @@ const formStyles = tv({
         sectionDescription: 'px-trim-md',
         fieldSet: 'flex flex-col gap-trim-md pt-trim-md',
         field: `grid grid-cols-2 [grid-template-areas:'header_control''description_description'] gap-x-trim-lg gap-y-0 p-trim-md border border-separator rounded-sm`,
-        labelContainer: '[grid-area:header]',
-        labelText: 'pb-trim-md text-base-fg text-lg',
         description: '[grid-area:description] text-base text-description',
         control: '[grid-area:control] flex items-center justify-end',
         // The error block is outside the named areas; span it full-width so `layout='full'` settings forms don't misplace it.
