+++ packages/ui/react-ui-form/src/components/Form/FormField/index.ts
@@ -6,5 +6,6 @@ export * from './FieldContainer';
 export * from './FieldHeader';
 export * from './FormField';
 export * from './fields';
+export * from './FormFieldLabel';
 export * from './FormFieldWrapper';
 export * from './presentation';
